Give the compiler optimization levels, and three passes to put in them
`--debug` and `--release` were one switch wearing two hats: the only thing
release did was require exhaustive `match`, which is a diagnostic and not an
optimisation at all. Split them into two independent axes — `OptLevel` and
`Strictness` — and make the profiles names for a bundle of the two rather than
things the compiler knows about.

Three layers now decide what a build does, each overriding the last: the
profile's built-in meaning, the package's `[profile.<name>]` section, and flags
on the command line (`-O`, `--strict`, `--lenient`). A key the manifest does not
mention is not overridden, so a package that only wants `opt-level = 2` in debug
builds says exactly that.

Then three back-end passes that trade nothing, and are therefore always on:

* an instruction wanting a window of consecutive registers reads the environment
  where it already sits when the environment already has that shape, which it
  usually does. Most of what codegen emitted was `move rN+1 <- rN` before every
  `closure` and every `invoke`.
* a literal operand folds into the instruction that uses it, so it takes no
  register and is not carried through the moves at the next jump.
* a comparison fuses into the branch that tests it, so the boolean never exists.

A tail loop went from 12 instructions an iteration to 6. And one pass that does
trade — `match` compiled to a decision tree — gated at `-O2`.

Measured on one machine, against the same build before the change: ack 3 3
244 -> 238 us, naiveFib 20 1.51 -> 1.35 ms, sumList 165 -> 160 us, buildVector
1.39 -> 1.22 ms, joinNumbers 848 -> 789 us.

One thing found the slow way and worth recording: widening `run_prim`'s operands
from `[u8; 3]` to `[usize; 3]` — so a fused branch could address a scratch slot —
cost about 4% across every benchmark, because the wider array is passed on the
stack rather than in a register. It is the innermost call in the machine. The
fix is a register the compiler does not use (`vm::TEMP`), which is sound only
because a fused branch's primitive is always a comparison and no comparison
allocates; `Prim::compares` is the list, codegen checks against it, and
`a_comparison_allocates_nothing` checks the list is still telling the truth.

Tests: every case in `differential.rs` now runs at all three levels rather than
the default, since the levels differ in the lowering and those are the only
cases that put each construct through both shapes one at a time. `opt.rs` adds
19 more, each requiring one answer from the CEK machine and the VM at O0, O1 and
O2 — the folding and fusing ones are mostly about `1 - n` and `n - 1` being
different numbers.

+++ rts/src/prims.rs
@@ -127,9 +127,15 @@ impl Vm<'_> {
     /// Registers rather than values: a primitive that allocates has to re-read
     /// its arguments after making room, because the collector will have moved
     /// them. See the discipline at the top of this file.
+    ///
+    /// Register *numbers* rather than indices, and that is not incidental:
+    /// `[u8; 3]` is passed in a register and `[usize; 3]` is passed on the
+    /// stack. This is the innermost call in the machine — every arithmetic
+    /// operation the program performs — and widening it cost about 4% across
+    /// every benchmark.
     pub(crate) fn run_prim(&mut self, p: Prim, srcs: [u8; 3], dst: u8) -> Result<(), Error> {
         use Prim::*;
-        let arg = |vm: &Vm, i: u8| vm.reg(srcs[i as usize]);
+        let arg = |vm: &Vm, i: usize| vm.reg(srcs[i]);
 
         let out = match p {
             // --- Int ------------------------------------------------------
